Enemy flip to player on MoveState
If enemy flips when touching a wall
it will go under the ground probably due to pivot. Temporal fix making enemy stop his movement when a wall is detected at a prudent distance. May change pivots in the future but this works for now just in case I add more enemies with diferent pivots and because I'm planning to use the wall detection mechanic for jumping anyways